Implement fix to handle weird host programs
We need to do something to handle the case where
the first token of a statement is something
that would be tokenized differently
in language vs shell mode.

We could tokenize in language mode and then  backtrack, but as far as I
can tell it will always work to just tokenize the first token
in shell mode always.